Fix megatron dry_run head count and fp16 dtype mismatch
dry_run used n_head=2 which isn't divisible by TP=4; bumped to n_head=4.
Cast model to fp16 after moving to CUDA since use_cpu_initialization
leaves params in float32, causing dtype mismatch in the pipeline schedule.

+++ Megatron.py
@@ -95,7 +95,7 @@ def _build_model(model_cfg: dict, seq_len: int) -> GPTModel:
         post_process=mpu.is_pipeline_last_stage(),
     )
 
-    return model.cuda()
+    return model.cuda().half()
 
 
 # ── Data iterator ─────────────────────────────────────────────────────────────
